Renamed View folder to Views to eliminate namespace conflict
