+++ DM/DM/Views/RecentMatchesPage.xaml
@@ -6,7 +6,7 @@
              xmlns:local="clr-namespace:DM.Help"
              xmlns:ffimageloading="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms"
              mc:Ignorable="d"
-             x:Class="DM.View.RecentMatchesPage">
+             x:Class="DM.Views.RecentMatchesPage">
     <ContentPage.Resources>
         <ResourceDictionary>
             <local:TeamWonConverter x:Key="teamWon"/>
