+++ DM/DM/Views/WelcomePage.xaml
@@ -6,7 +6,7 @@
              xmlns:ffimageloading="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms"
              xmlns:local="clr-namespace:DM.Help"
              mc:Ignorable="d"
-             x:Class="DM.View.WelcomePage">
+             x:Class="DM.Views.WelcomePage">
     <ContentPage.Resources>
         <ResourceDictionary>
             <local:RankConverter x:Key="rankConverter"/>
