+++ DM/DM/Views/WelcomePage.xaml.cs
@@ -3,7 +3,7 @@ using DM.ViewModels;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
-namespace DM.View
+namespace DM.Views
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class WelcomePage : ContentPage
